+++ src/egon/data/airflow/dags/pipeline.py
@@ -19,6 +19,7 @@ import egon.data.importing.etrago as etrago
 import egon.data.importing.mastr as mastr
 import egon.data.processing.substation as substation
 import egon.data.importing.re_potential_areas as re_potential_areas
+import egon.data.importing.heat_demand_data as import_hd
 
 # Prepare connection to db for operators
 airflow_db_connection()
@@ -193,7 +194,6 @@ with airflow.DAG(
     substation_functions >> hvmv_substation_extraction
     substation_functions >> ehv_substation_extraction
 
-
     # Import potential areas for wind onshore and ground-mounted PV
     download_re_potential_areas = PythonOperator(
         task_id="download_re_potential_area_data",
@@ -210,3 +210,11 @@ with airflow.DAG(
     )
     setup >> download_re_potential_areas >> create_re_potential_areas_tables
     create_re_potential_areas_tables >> insert_re_potential_areas
+
+    # Future heat demand calculation based on Peta5_0_1 data
+    heat_demand_import = PythonOperator(
+        task_id="import-heat-demand",
+        python_callable=import_hd.future_heat_demand_data_import
+    )
+    vg250_clean_and_prepare >> heat_demand_import
+    population_import >> heat_demand_import